+++ edit-me/config/resumeConfig.ts
@@ -1,9 +1,17 @@
-import { AccentColors, NeutralColors, ResumeConfig, Theme } from './Config';
+import {
+  AccentColors,
+  NeutralColors,
+  ResumeConfig,
+  Theme,
+  ThemeSetting,
+} from './Config';
 
 const config: ResumeConfig = {
   accentColor: AccentColors.Blue,
   neutralColor: NeutralColors.Mauve,
+  appTheme: ThemeSetting.Dark,
   imageTheme: Theme.Light,
+  pdfTheme: Theme.Light,
 };
 
 export default config;